+++ CHANGELOG.md
@@ -2,6 +2,18 @@ Changelog
 =========
 
 ## HEAD
+*   support only go 1.13, due to use of new error wrapping semantics
+*   improved client connection early abort handling
+*   improved max response side handling -- only read MaxSize KB from any
+    upstream server. Note: This may result in partial responses to clients
+    for chunked encoding requests that are longer than MaxSize, as there is
+    no way to signal the client other than closing the connection.
+*   change default --max-size to 0
+    previously chunked encoding responses bypassed size restrictions. to
+    avoid unexpected failures (preserve backwards compate in this
+    regard), set max-size to 0 by default moving forward.
+    previous default was 5mb (`--max-size=5120`)
+
 
 ## v2.0.1 2019-09-12
 *   Slightly optimize some structure layouts to reduce memory overhead.